refactor: extract calendar patch into hidecalendar package
Move hideCalendarButtonPatch and its five fingerprints out of the
chatheaderbuttons package into a dedicated hidecalendar package, grouping
by feature (calendar) rather than by surface. Rename the user-facing patch
from "Hide LINE Calendar" to "Hide calendar buttons".

chatheaderbuttons now holds only the community button; its shared enum/list
descriptor consts stay there for CommunityButtonFingerprint.
